ingress-proxy-configurer/app/requirements.txt
appdirs==1.4.3
backports.ssl-match-hostname==3.5.0.1
click==6.7
docker-py==1.10.6
docker-pycreds==0.2.1
Flask==0.12.1
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
requests==2.14.2
six==1.10.0
websocket-client==0.40.0
Werkzeug==0.12.1
